back/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
django-cors-headers==4.6.0
numpy==2.0.2
gunicorn==23.0.0
psycopg==3.2.3
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
ruff==0.9.3
django-filter==24.3
django-silk==5.3.2
grpcio==1.70.0
grpcio-tools==1.70.0
